requirements.txt
# Requirements for SMPC Collaborative Data Analysis
# Core cryptographic library
pycryptodome==3.19.0

# Optional: For enhanced logging and formatting
colorlog==6.8.0

# Development and testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1